rename: (폴더명: childinfo, record), (컴포넌트명: RegisterChildInfo, RegisterChildRecord) 로 변경
