+++ src/components/child/ChildCard.tsx
@@ -67,7 +67,7 @@ export const ChildCard = ({ child }: ChildCardProps) => {
                   </p>
 
                   {/* 수정하기 */}
-                  <Link key={child.id} href={`/child/${child.id}/firststep`}>
+                  <Link key={child.id} href={`/child/${child.id}/childinfo`}>
                     <div className="flex justify-center items-center self-stretch flex-grow-0 flex-shrink-0 h-8 relative gap-[8px] p-[7px] rounded-[7px] bg-gray-30 w-[180px]">
                       <button className="flex-grow-0 flex-shrink-0 text-xs font-medium text-center text-gray-700">
                         수정하기
